requirements.txt
pytest-bdd
requests-mock
requests
pydantic
pytest
python-dotenv
